backend/requirements.txt
loguru
supabase
python-dotenv
fastapi
python-jose
slowapi
aiofiles
redis
pycryptodome
httpx
pydantic
uvicorn